requirements.txt
flask==3.0.2
ir_datasets==0.5.7
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
rank-bm25==0.2.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.5.1
textstat==0.7.3
